refactor: move menu button to same location as menu
