+++ src/components/buttons/index.tsx
@@ -1,5 +1,4 @@
 export * from "./NeutralButton";
 export * from "./PrimaryButton";
 export * from "./NewCategoryButton";
-export * from "./MenuButton";
 export * from "./MoreButton";
